pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "configmanager"
version = "0.0.1"
authors = [
  { name="yamause"},
]
description = "Config manager for python"
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyYAML",
    "pydantic",
    "python-dotenv"
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[tool.setuptools]
package-dir = {"" = "src"}
